+++ packages/editor/src/plugins/email-theming/themes.ts
@@ -96,7 +96,7 @@ const THEME_BASIC: PanelGroup[] = [
       {
         label: 'Align',
         type: 'select',
-        value: 'left',
+        value: 'center',
         options: {
           left: 'Left',
           center: 'Center',
@@ -448,9 +448,17 @@ const THEME_MINIMAL: PanelGroup[] = [
         classReference: 'body',
       },
       {
+        label: 'Margin',
+        type: 'number',
+        value: 0,
+        unit: 'px',
+        prop: 'margin',
+        classReference: 'body',
+      },
+      {
         label: 'Padding Top',
         type: 'number',
-        value: undefined,
+        value: 0,
         unit: 'px',
         prop: 'paddingTop',
         classReference: 'body',
@@ -458,7 +466,7 @@ const THEME_MINIMAL: PanelGroup[] = [
       {
         label: 'Padding Right',
         type: 'number',
-        value: undefined,
+        value: 0,
         unit: 'px',
         prop: 'paddingRight',
         classReference: 'body',
@@ -466,7 +474,7 @@ const THEME_MINIMAL: PanelGroup[] = [
       {
         label: 'Padding Bottom',
         type: 'number',
-        value: undefined,
+        value: 0,
         unit: 'px',
         prop: 'paddingBottom',
         classReference: 'body',
@@ -474,7 +482,7 @@ const THEME_MINIMAL: PanelGroup[] = [
       {
         label: 'Padding Left',
         type: 'number',
-        value: undefined,
+        value: 0,
         unit: 'px',
         prop: 'paddingLeft',
         classReference: 'body',
@@ -1096,6 +1104,13 @@ export const SUPPORTED_CSS_PROPERTIES: SupportedCssProperties = {
     type: 'color',
     defaultValue: '#000000',
     category: 'appearance',
+  },
+  margin: {
+    label: 'Margin',
+    type: 'number',
+    unit: 'px',
+    defaultValue: 0,
+    category: 'layout',
   },
   padding: {
     label: 'Padding',